Move some python functions to helper modules
